Utest_fw: properly print integer-typed values
The old implementation just lead to an endless loop.

+++ src/test/utest/framework/utest_fw.cpp
@@ -511,7 +511,7 @@ template <typename T> inline
 void
 utest_format_print_value(T const &val)
 {
-  utest_format_print_value(cxx::int_value<T>(val));
+  utest_format_print_value((unsigned long)cxx::int_value<T>(val));
 }
 
 inline
